requirements.txt
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.1
sqlparse==0.3.0
